backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
supabase
redis
celery
python-dotenv 
greenlet
psycopg2
